fix(install): harden trusted native evidence verification
